(Fixes #517) Fix header option for SplitMergeFilesActivity
Currently, if there is only 1 input file and partitions are not required the
header option is ignored.
This change fixes it so under these circumstances the header will be added to
the output file. It also retains the original functionality of just moving
the input file to the output location if the header is not used.