Fixes ##602
Without this patch, driver programs doesn't have any knowledge of created pipeline details. Thus reducing the capability of the driver programs to manage and monitor created pipelines.

This patch helps in providing hook for the driver program to get pipelineId, name & status after pipeline creation as a call back, so drivers have the freedom to run custom functionality (if needed).